Massive refactoring on #include statements + changed widths and heights types from unsigned integers to floats

+++ OpenNeighborhood/src/Elements/Element.h
@@ -18,16 +18,16 @@ public:
 
 	inline const std::string& GetLabel() const { return m_Label; }
 	inline const std::string& GetTextureName() const { return m_TextureName; }
-	inline uint32_t GetWidth() const { return m_Width; }
-	inline uint32_t GetHeight() const { return m_Height; }
+	inline float GetWidth() const { return m_Width; }
+	inline float GetHeight() const { return m_Height; }
 
 	using EventCallbackFn = std::function<void(Event&)>;
 	inline void SetEventCallback(const EventCallbackFn& callback) { m_EventCallback = callback; }
 protected:
 	std::string m_Label;
 	std::string m_TextureName;
-	uint32_t m_Width;
-	uint32_t m_Height;
+	float m_Width;
+	float m_Height;
 	float m_Padding = 10.0f;
 	EventCallbackFn m_EventCallback;
