+++ OpenNeighborhood/src/Panels/MainPanel.cpp
@@ -13,7 +13,7 @@ void MainPanel::OnRender()
 	;
 
 	ImGui::SetNextWindowPos(ImVec2(0.0f, 0.0f));
-	ImGui::SetNextWindowSize(ImVec2((float)m_WindowWidth, (float)m_WindowHeight));
+	ImGui::SetNextWindowSize(ImVec2(m_WindowWidth, m_WindowHeight));
 	ImGui::PushStyleVar(ImGuiStyleVar_WindowRounding, 0.0f);
 
 	ImGui::Begin("Main Window", nullptr, windowFlags);
